==> research/03_model_training.ipynb <==
import os


os.chdir('../')

from dataclasses import dataclass
from pathlib import Path
@dataclass(frozen=True)
class PrepareTrainingConfig:
    root_dir:Path
    trained_model_path:Path
    updated_base_model:Path
    training_data:Path
    params_epochs:int
    params_batch_size:int
    params_image_size:list
    

    
    


from AI_Vs_Human_Geoguess.constants import *
from AI_Vs_Human_Geoguess.utils.common import read_yaml,create_directories
import tensorflow as tf



    
class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH):

        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)

        create_directories([self.config.artifacts_root])


    
    def get_training_config(self) -> PrepareTrainingConfig:
        training=self.config.training
        prepare_base_model=self.config.prepare_base_model
        params=self.params
        training_data=os.path.join(self.config.data_ingestion.unzip_dir,'dataset')
        create_directories([Path(training.root_dir)])

        training_config=PrepareTrainingConfig(
            root_dir=Path(training.root_dir),
            trained_model_path=Path(training.trained_model_path),
            updated_base_model=Path(prepare_base_model.updated_base_model_path),
            training_data=Path(training_data),
            params_epochs=params.EPOCHS,
            params_batch_size=params.BATCH_SIZE,
            params_image_size=params.IMAGE_SIZE
        )
        return training_config


import os
import zipfile
import gdown
from AI_Vs_Human_Geoguess import logger
from AI_Vs_Human_Geoguess.utils.common import get_size
import pandas as pd
import tensorflow as tf


class Training:
    def __init__(self,config:PrepareTrainingConfig):
        self.config=config
    def get_base_model(self):
        self.model=tf.keras.models.load_model(
            self.config.updated_base_model
        ) 
        self.data=pd.read_csv(os.path.join(self.config.training_data,'coords.csv'))
    def train_validation_generator(self):
        datagenerator_kwargs=dict(
            rescale=1/255,
            validation_split=0.2
        )
        dataflow_kwargs=dict(
            dataframe=pd.read_csv(os.path.join(self.config.training_data,'coords.csv')),
            directory=self.config.training_data,
            x_col='image',
             y_col=['normalized_lat', 'normalized_long'],  # Specify that this subset is for training
            target_size=(180, 180),
            batch_size=32,
            class_mode='raw'
            
        )
        valid_datagenerator=tf.keras.preprocessing.image.ImageDataGenerator(
            **datagenerator_kwargs
        )
        self.valid_generator=valid_datagenerator.flow_from_dataframe(
            
            subset='validation',
            **dataflow_kwargs  # Specify that this subset is for training
            

        )

        training_generator=valid_datagenerator
        self.train_generator=training_generator.flow_from_dataframe(
            subset='training',
            ** dataflow_kwargs
        )
    def train(self):
        self.steps_per_epoch = self.train_generator.samples // self.train_generator.batch_size
        self.validation_steps = self.valid_generator.samples // self.valid_generator.batch_size
        self.model.summary()
        self.model.fit(
            self.train_generator,
            epochs=self.config.params_epochs,
            steps_per_epoch=self.steps_per_epoch,
            validation_steps=self.validation_steps,
            validation_data=self.valid_generator,
            callbacks=tf.keras.callbacks.ModelCheckpoint('content/best_model/model.keras', save_best_only=True, monitor='val_loss', mode='min')

        )

        self.save_model(
            path=self.config.trained_model_path,
            model=self.model
        )
    




try:
    config=ConfigurationManager()
    training_config=config.get_training_config()
    training=Training(config=training_config)
    training.get_base_model()
    training.train_validation_generator()
    training.train()
except Exception as e:
    raise e
